Big changes on submodules projects
