--- Banking_Functionallity/src/main/java/dev/adrylat/bankingfunctionality/card/MyException.java
@@ -1,20 +1,0 @@
-package dev.adrylat.bankingfunctionality.card;
-
-import java.util.List;
-
-public class MyException extends Exception {
-
-    private  List<String> errorList;
-
-    public MyException(List<String> errorList) {
-        this.errorList = errorList;
-    }
-
-    public List<String> getErrorList() {
-        return errorList;
-    }
-
-    public  void showErrorsLog() {
-        errorList.forEach(System.err::println);
-    }
-}
