+++ banking_functionality/src/main/java/dev/adrylat/bankingfunctionality/card/exceptions/CardValidatorException.java
@@ -1,0 +1,17 @@
+package dev.adrylat.bankingfunctionality.card.exceptions;
+
+import java.util.List;
+
+public class CardValidatorException extends Exception {
+
+    private  List<String> errorList;
+
+    public CardValidatorException(List<String> errorList) {
+        this.errorList = errorList;
+    }
+
+    public List<String> getErrorList() {
+        return errorList;
+    }
+
+}
