+++ banking_functionality/src/main/java/dev/adrylat/bankingfunctionality/card/validators/CardValidator.java
@@ -1,0 +1,8 @@
+package dev.adrylat.bankingfunctionality.card.validators;
+
+import dev.adrylat.bankingfunctionality.card.exceptions.CardValidatorException;
+
+public interface CardValidator {
+
+   boolean validate(String input) throws CardValidatorException;
+}
