+++ banking_functionality/src/main/java/dev/adrylat/bankingfunctionality/mortgagecalculator/CustomerData.java
