+++ pom.xml
@@ -7,6 +7,10 @@
     <artifactId>banking_utilities</artifactId>
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
+    <modules>
+        <module>banking_functionality</module>
+        <module>dialogs</module>
+    </modules>
     <build>
         <plugins>
             <plugin>
@@ -18,7 +22,7 @@
                         <manifest>
                             <addClasspath>true</addClasspath>
                             <classpathPrefix>lib/</classpathPrefix>
-                            <mainClass>dev.andrylat.bankingutilities.MainApplication</mainClass>
+                            <mainClass>dev.andrylat.dialogs.MainApplication</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
@@ -33,10 +37,6 @@
             </plugin>
         </plugins>
     </build>
-    <modules>
-        <module>Banking_Functionallity</module>
-        <module>Dialogs</module>
-    </modules>
     <dependencies>
         <!-- https://mvnrepository.com/artifact/org.apache.maven.plugins/maven-failsafe-plugin -->
         <dependency>
@@ -55,18 +55,6 @@
             <groupId>org.apache.maven.surefire</groupId>
             <artifactId>surefire-api</artifactId>
             <version>3.0.0-M5</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>dev.andrylat</groupId>
-            <artifactId>Banking_Functionallity</artifactId>
-            <version>1.0-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>dev.andrylat</groupId>
-            <artifactId>Dialogs</artifactId>
-            <version>1.0-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
